+++ CHANGES.rst
@@ -14,11 +14,13 @@ Improvements:
 - Calling getall with soap backend will always return list of object, even if SOAP response has one element (`#1486`_)
 - Introduce `base64()` prepare function that decodes base64 string (`#1486`_)
 - `base64()` prepare function that decodes base64 binary now does this outside dask dataframes. Also, raises `InvalidBase64String` for invalid base64 (`#1486`_)
+- XSD manifests now support `xs:key` element to define primary keys. Keys are extracted from `xs:selector` and `xs:field` elements and mapped to the `ref` column in generated manifests. Supports both simple and composite keys. (`#1472`_)
 
   .. _#605: https://github.com/atviriduomenys/spinta/issues/605
   .. _#675: https://github.com/atviriduomenys/spinta/issues/675
   .. _#1461: https://github.com/atviriduomenys/spinta/issues/1461
   .. _#1462: https://github.com/atviriduomenys/spinta/issues/1462
+  .. _#1472: https://github.com/atviriduomenys/spinta/issues/1472
   .. _#1486: https://github.com/atviriduomenys/spinta/issues/1486
 
 Other: